requirements.txt
absl-py==1.1.0
ale-py==0.7.5
astunparse==1.6.3
atari-py==0.3.0
attrs==21.4.0
bsuite==0.3.5
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
chex==0.1.3
cloudpickle==2.1.0
cycler==0.11.0
decorator==5.1.1
descartes==1.1.0
dill==0.3.5.1
distrax==0.1.2
dm-control==0.0.364896371
dm-env==1.5
dm-haiku==0.0.6
dm-launchpad==0.5.2
dm-reverb==0.7.2
dm-sonnet==2.0.0
dm-tree==0.1.7
flatbuffers==2.0
flax==0.5.0
fonttools==4.33.3
frozendict==2.3.2
future==0.18.2
gast==0.5.3
glfw==2.5.3
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grpcio==1.46.3
gym==0.24.1
gym-notices==0.0.7
h5py==3.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
jmp==0.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
labmaze==1.0.5
libclang==14.0.1
lxml==4.9.0
Markdown==3.3.7
matplotlib==3.5.2
mizani==0.7.4
mock==4.0.3
msgpack==1.0.4
networkx==2.8.3
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
palettable==3.3.0
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.1
pip==22.1.2
plotnine==0.8.0
pluggy==1.0.0
portpicker==1.5.2
promise==2.3
protobuf==3.20.1
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.0
requests-oauthlib==1.3.1
rlax==0.1.2
rlds==0.1.4
rsa==4.8
scikit-image==0.19.2
scipy==1.8.1
setuptools==62.3.3
six==1.16.0
statsmodels==0.13.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
tensorflow-probability==0.15.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.5.4
tomli==2.0.1
toolz==0.11.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.0
